requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
clickhouse-driver==0.2.6
python-dotenv==1.0.1
